chore: move parseClassFromString to scripts.js
